post timestamp 잘못 기록되는 오류 해결
문자열이 아닌 정수 형식으로 저장되도록 변경

issue-049

+++ app/src/main/java/hello/kwfriends/ui/screens/findGathering/findGatheringCardList.kt
@@ -50,7 +50,7 @@ fun GathergingListItem(
                     ) {
                         FlowRow(verticalArrangement = Arrangement.Center) {
                             Text(
-                                text = SimpleDateFormat("MM/dd HH:mm", Locale.getDefault()).format(postDetail.timestamp.toLong()),
+                                text = SimpleDateFormat("MM/dd HH:mm", Locale.getDefault()).format(postDetail.timestamp),
                                 maxLines = 1,
                                 style = MaterialTheme.typography.bodySmall,
                                 fontFamily = FontFamily.Default,
