+++ app/src/main/java/hello/kwfriends/ui/screens/postInfo/PostInfoScreen.kt
@@ -141,7 +141,7 @@ fun PostInfoScreen(
                         fontWeight = FontWeight(500)
                     )
                     Text(
-                        text = SimpleDateFormat("MM/dd HH:mm", Locale.getDefault()).format(postDetail.timestamp.toLong()),
+                        text = SimpleDateFormat("MM/dd HH:mm", Locale.getDefault()).format(postDetail.timestamp),
                         maxLines = 1,
                         style = MaterialTheme.typography.bodySmall,
                         fontFamily = FontFamily.Default,
